Clean-up

+++ core-it/src/main/kotlin/bar/examples/it/spring/aiclassification/allornone/AllOrNoneApplication.kt
@@ -21,7 +21,7 @@ open class AllOrNoneApplication {
 
         var b : String? = null,
 
-        var c : Long? = null,
+        var c : Double? = null,
 
         var d : ACAllOrNoneEnum? = null,
